move redirect_resolver to LspResolver
